backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
youtube-transcript-api==0.6.2
openai==1.12.0
python-dotenv==1.0.0
